Refactoring, bug fixes, restyling existing elements
Refactored styling of elements into a separate method. (Call it and it
will give back proper info for $text in li style="$text")

Added restyling of existing elements so that when you load the
spreadsheet it restyles the elements already on screen. This was much
simpler than I thought it'd be. (Although time will tell)

Fixed a bug with visited companies sheet not working properly.

Fixed a bug with crunchbases scrolling boundaries not working properly.

Refactored magic numbers into variables instead.